requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.3 ; python_version >= '3.6'
aiosignal==1.3.1 ; python_version >= '3.7'
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==23.1.0 ; python_version >= '3.7'
certifi==2023.5.7 ; python_version >= '3.6'
charset-normalizer==2.0.12 ; python_full_version >= '3.5.0'
chess.com==2.0.6
discord-pretty-help==2.0.7
discord.py==2.3.1
frozenlist==1.3.3 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
multidict==6.0.4 ; python_version >= '3.7'
python-dotenv==1.0.0
python-lichess==0.10
requests==2.28.0
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
urllib3==1.26.16 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
yarl==1.9.2 ; python_version >= '3.7'